pyproject.toml
[project]
name = "mlops-101"
version = "0.1.0"
description = "MLOps 101: For Students by Students"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gcsfs>=2025.2.0",
    "google-cloud-storage>=3.0.0",
    "ipykernel>=6.29.5",
    "mlflow==2.20.1",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "pydantic>=2.10.6",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "requests>=2.32.3",
    "requests-mock>=1.12.1",
    "scikit-learn>=1.6.1",
]

[build-system]
requires = [
  "setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where =["src"]
include = ["make_data*"]

[tool.ruff]
line-length = 120
select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
